\name{MDSGUI-package}
\alias{MDSGUI-package}
\alias{MDSGUI}
\docType{package}
\title{
An R package providing access to the MDS-GUI
}
\description{
A graphical user interface (GUI) for performing Multidimensional Scaling applications and interactively analysing the results all within the GUI environment. The MDS-GUI provides means of performing Classical Scaling, Least Squares Scaling, Metric SMACOF, Non-Metric SMACOF, Kruskals Analysis and Sammon Mapping with animated optimisation.
}
\details{-
\tabular{ll}{
Package: \tab MDSGUI\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-08-28\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Andrew Timm and Sugnet Gardner-Lubbe

Maintainer: Andrew Timm <timmand@gmail.com>
}
\note{                  The GUI was developed in R with the tcltk package. For the best results it is recommended that R-2.13.0 be used. Also, package version 2_1.1-5 of tcltk2 and 0.0-23 of tkrplot produce the best results. 
}
\references{
All MDS is based on the theory covered in "Multidimensional Scaling: Second Edition" by Cox, T.G. and Cox, M.A. (2001) and "Modern Multidimensional Scaling: Theory and Applications Second Edition" by Borg, I. and Groenen, P.J.F. (2005).
}

\keyword{ package }

